+++ src/Expression/Level1.php
@@ -10,7 +10,7 @@ use Innmind\UrlTemplate\{
 use Innmind\Immutable\{
     Map,
     Str,
-    Maybe,
+    Attempt,
 };
 
 /**
@@ -31,13 +31,13 @@ final class Level1 implements Expression
     /**
      * @psalm-pure
      *
-     * @return Maybe<self>
+     * @return Attempt<self>
      */
-    public static function of(Str $string): Maybe
+    public static function of(Str $string): Attempt
     {
-        return Name::one($string, Expansion::simple)->map(
-            static fn($name) => new self($name),
-        );
+        return Name::one($string, Expansion::simple)
+            ->map(static fn($name) => new self($name))
+            ->attempt(static fn() => new \LogicException('Cannot parse level 1'));
     }
 
     /**