+++ src/Expression/Level3/Path.php
@@ -13,7 +13,7 @@ use Innmind\Immutable\{
     Map,
     Sequence,
     Str,
-    Maybe,
+    Attempt,
 };
 
 /**
@@ -40,13 +40,13 @@ final class Path implements Expression
     /**
      * @psalm-pure
      *
-     * @return Maybe<self>
+     * @return Attempt<self>
      */
-    public static function of(Str $string): Maybe
+    public static function of(Str $string): Attempt
     {
-        return Name::many($string, Expansion::path)->map(
-            static fn($names) => new self($names),
-        );
+        return Name::many($string, Expansion::path)
+            ->map(static fn($names) => new self($names))
+            ->attempt(static fn() => new \LogicException('Cannot parse level 3'));
     }
 
     #[\Override]