+++ src/Expression/Level3/QueryContinuation.php
@@ -12,7 +12,7 @@ use Innmind\Immutable\{
     Map,
     Sequence,
     Str,
-    Maybe,
+    Attempt,
 };
 
 /**
@@ -34,13 +34,13 @@ final class QueryContinuation implements Expression
     /**
      * @psalm-pure
      *
-     * @return Maybe<self>
+     * @return Attempt<self>
      */
-    public static function of(Str $string): Maybe
+    public static function of(Str $string): Attempt
     {
-        return Name::many($string, Expansion::queryContinuation)->map(
-            static fn($names) => new self($names),
-        );
+        return Name::many($string, Expansion::queryContinuation)
+            ->map(static fn($names) => new self($names))
+            ->attempt(static fn() => new \LogicException('Cannot parse level 3'));
     }
 
     public static function named(Name $name): self